refactored tpw package to data, added missing comments
