+++ src/main/java/cz/uhl1k/typewriter/data/Data.java
@@ -16,7 +16,7 @@
   along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-package cz.uhl1k.typewriter.tpw;
+package cz.uhl1k.typewriter.data;
 
 import cz.uhl1k.typewriter.exceptions.NoFileSpecifiedException;
 import cz.uhl1k.typewriter.model.Book;
