+++ src/main/java/cz/uhl1k/typewriter/data/TpwFileHandler.java
@@ -16,7 +16,7 @@
   along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-package cz.uhl1k.typewriter.tpw;
+package cz.uhl1k.typewriter.data;
 
 import cz.uhl1k.typewriter.model.Book;
 import cz.uhl1k.typewriter.model.Chapter;
