+++ src/main/java/cz/uhl1k/typewriter/data/package-info.java
@@ -1,0 +1,2 @@
+/** Package containing classes for manipulation with files and data. */
+package cz.uhl1k.typewriter.data;
