+++ src/main/java/cz/uhl1k/typewriter/exceptions/SettingsNotSavedException.java
@@ -1,3 +1,4 @@
 package cz.uhl1k.typewriter.exceptions;
 
+/** Thrown when settings were not saved. */
 public class SettingsNotSavedException extends Exception {}
